requirements.txt
Flask==3.1.3
gunicorn==22.0.0
scikit-learn
pandas
numpy
blinker==1.9.0
click==8.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
Flask-CORS==6.0.2
setuptools
python-dotenv==1.0.0
joblib
requests==2.31.0